Action Table
The 'faithful rendition' plan has gone straight out the window, since I
realized that there were two ArrayLists embedded in the action_table.

So I'm just making the best of it.  The prospect of leaving methods on
the exterior of the data structures they work on was also pretty gross
to me, so yeah.  I'm leaving line-number breadcrumbs so I can quickly
find the original of this or that, and just charging ahead.