Configlist: port the easy stuff
This code uses a lot of nullables.  It does so in a valid manner,
most of them are linked lists, and that's how those work.  I don't
mind the nullability, but I really mind that it isn't a part of the
type system, so I need to deduce it.

Anyway, that's another chunk out of the way.  The nature of this task
is such that I can't meaningfully test things until I've effectively
completed the port itself.  Small favors, this doesn't depend on
anything but the standard library, so I won't have to fight too hard
to get it into a debugger.

Slightly larger favor, I can add identical prints to lemon.c and to
lemon.zig, and just diff them.