Parse decl keyword state
A shining example of how Zig can improve on idiomatic C at lower cost,
thanks to comptime magic and the real data structures it enables.

Instead of a chain of strcmps, which the compiler cannot reasonably
optimize into anything more efficient, we use a comptime-known
StaticStringMap to convert known states into an enum, and switch on
that.  This compares lengths and uses a binary search, minimizing
the amount of byte comparison which is even necessary to a worst-case
of the number of strings of equal length.

Also, we get an optional enum, so I can make the null case a switch
break after issuing either of the two kinds of error.  It's always
a bit nicer to dispose of error cases rapidly, and in this case it
also saves us a level of nesting, and those are never free.

It would be absurd to claim that this routine was in any sense in
need of optimization.  But the code _is_ faster, and it's also better,
for instance, when I change the decls for Zitron (which I will) the
exhaustive switch will make sure I handle each new option as it is
added.

I could have used a comptime loop to populate the list with tagnames,
but honestly, copy-pasting and using a Vim macro to do the text
transform was almost as fast.  For Zitron I might do it that way so
there's one source of truth instead of two.