Scanner
This is a port of the core logic of the Parse function.  True to form,
Lemon tokenizes the file buffer, passing each token to the parser as it
is created.  I'm thankful that the bootstrap urge was ignored, meaning
Lemon wasn't rewritten as a lemon.y file, which would require a compiled
lemon to compile a lemon from it.  Given that the original files and the
aagen tool are lost to time, that would have been a fragile thing to do.

That's as far as I'm likely to get today.  I need to back fill some of
the parts, like the symbol intern table, make things initializable, and
so on, but I'm almost ready to run lemon.zig on pikchr.y, with print
statements that will start with emitting tokens separated by a pretty
UTF-8 sigil which isn't found in either codebase.  That will at least
demonstrate that it's sliced up right, we'll see what happens next.