Various intern structures
Lemon.c is about 6000 lines, and this commit gets us to 1800.  I
expect lemon.zig to be longer, given comments and Zig sometimes taking
more lines to do the same thing, perhaps an additional thousand, but
we'll see.

Still! Substantial progress. Right on the verge of getting a bunch of
this code to compile...