Gameplan entered
The idea here is simple: we're going to make lemon.c and lemon.zig
very loud, one step at a time.  They're supposed to do the same
things in the same order, so we make them say the things they do
in the same way.  Then we run them both and diff the outputs.

I'm going to treat parsing as a given for now, even though there's
simply gotta be at least one error in translation.  It's better for
momentum to start tackling the steps, and set up the scripts which
auto-run both players, then I can add stuff (back) to the parse
stage if that proves necessary pretty easily.