Sort symbols and rules
The ability to cross-print between lemon.c and lemon.zig begins to
pay off.  I spotted a scanning error in my code, and confirmed that
my comparison function and his give the same order for Pikchr.

The rule sort is a bit more of a story: I was able to notice that
it can be done with fewer iterations and no sorting at all, and that
also prints identically between the programs, so that feels pretty
good.  I still have to port that merge sort pretty soon though.