Begin entry point of FindStates
This procedure calls in to a substantial fraction of the compiling
portion of Lemon; only the main block of FindStates is ported here,
and untested at that.

Thus far, I've gotten along nicely with eyeball-grepping of matched
print statements between lemon.c and lemon.zig, using only Pikchr as
a template.  My intuition is telling me that I can save pain and
headache down the line, by taking the time out to collect a larger
set of `.y` files, such as SQlite itself, and write a little script
which runs both lemons against every such template, diverting all
stderr printing into paired files, then diffing those files at the
conclusion.  The actual output files will be ignored for now, but
as we get to generating our own, we can diff these as well.

The result can only be two versions of a program, which exhibit an
identical internal state at critical regions of the algorithms, then
emit byte-identical text, for every extant input to Lemon which I'm
able to dredge up.

The last stage will be to write 'trap files', which induce both Lemons
to emit every reachable error message, and check that they each exit
with the same code.  That's where I declare victory in the porting
process, as I don't care much about, say, small differences in what
happens in an out-of-memory condition, and I know of no feasible way
to affirm that they behave identically in the face of unanticipated
inputs.