Coup counted, merge sort fixed for all inputs
I was extremely suspicious of that LISTSIZE-1 special case, after having
a sleep on it, I realized it was in fact wrong.  It's in the original, I
now have two genuine bug reports on DRH code, the other being the parser
for hexadecimal numbers in Pikchr giving malformed inputs a spurious yet
basically harmlesss value, instead of rejecting it.

I'm unlikely to ever score a point on Dr. Knuth, so this is as good as I
can get it.  One heckuva silver medal.

Details: https://sqlite.org/forum/forumpost/63750d717c