Wiring it all up, getting segfaults
I already told the leak detector to knock it off, and now we're getting
segfaults.  This could be worse, it's a new wave of reachable code, and
I haven't been especially disciplined with initializing everything, let
alone figuring out which pointers are optional and which are not.

My back is kinda tweaked, so I plan to take a break, but adding prints,
and seeing where it quits, should do an adequate job of showing me what
goes off the rails where.  Plus I can reuse it for correctness down the
road.