Un-segfaulted
In my quest to un-leak prior to punting that down the road, I had
made the configlist-builder destroy all my configs.  So that was
wrong.

The hash function also relies on overflow, not surprising in
retrospect, and easy to fix, Zig being what it is.  So we get to
the bottom of the slide now.

Next step is to figure out what we've created, and how to print it
in a consistent and useful way, so I can see how our version might
differ from the original.  And so we go.