Add CompressTables
Nearly at the turning point here.  One more function building the
intermediate representation of the parser, and then it's on to the
report section, which dumps the entire internal parser state in a
handy text file.  If that looks right then it's on to generating
the actual parser's source code.