Running start on ReportTable
I plan to pace myself on the next lap.  This commit introduces the
output routine, ReportTable, and implements some of the main helper
routines for it.

I intend to embed the default template, and things being what they
are I'm just planning to read the entire provided template into
memory when there is one.  Other than that it's a straight a shoot
as I can manage, given distinctions like slices vs. pointers which
I judged would make life easier in the long run.