Print stack union
This is a fiddly bit of code, because it builds up a hash table to
de-dupe strings, but the hash itself becomes load-bearing, appearing
in the output.  So I can't just pull one out of the stdlib and use
that instead.

Which isn't the fiddly part, a quick roll-your-own hash array is not
a challenging data structure.  It was converting all the C idioms
for string handling into something which makes sense in Zig.

What's striking, and I'll blog about this, is how much less work my
ported function ended up doing.  No need to allocate and free a
scratch buffer just to null-terminate a trimmed string, no need to
copy it from the buffer to the hash table, then free it in another
iteration.  Just allocate enough of a slice-of-slice to do the
job, borrow and non-destructively trim the strings, print them,
and free the hash table at the end.  With defer, so it still gets
freed if I have some kind of write error.  This is not a circumstance
where that matters, but you get it for free even when it does.