Destructor switch prints correctly
This was a bit of a steeplechase.

Not yet a clean diff: some number is skewed in the action table routine,
and I need to check if that's a problem.  It might just be a printing
problem, since the output is already slotted and looks great.

Amended, print error in lemon.c.