Add Zephir parser, passes conformance check
This found several incorrect logic branches which the SQLite and Pikchr
codes do not reach.  Hopefully that will peter out as I add more.

The comparison function for state sorting was an interesting one, since
the bug turned out to be that C uses qsort and Zig uses PDQ, the latter
will occasionally test an element against itself.