Add 'macro' system
Zig (fortunately) does not have macro replacement, so we need to use a
cheeky hashmap and do the replacements ourself.

Still a fair amount of grinding ahead, and frankly, how fast that goes
depends on how bullshit-tolerant I'm feeling.  The path forward is clear
and it's a bunch of tiny tweaks and translations.

One other thing: Zig complains if I free the template copy, and I'm just
not sure why that is. ¯\_(ツ)_/¯