Zig 0.15.1 compatibility
This one was fairly involved because we use Writers heavily.  That said,
the conversion was largely mechanical.

Also figured out why the template was leaking: I was bumping the slice
pointer, and therefore it pointed to the end of the memory allocation
when the defer executed.  No problem, just cache it.