Up to reduction rule generator
Changing the reduction rule generator is nontrivial, probably in a fun
way.  Then it should be downhill for awhile as I finish the template
translation.

Then there's a lot of deburring to do, then I need to write a sample
parser to exercise all this stuff.  Then it's ready for prime time,
well, kinda-sorta.