Tremendous ziggification progress
Completed a port of the Lemon manual to describe Zitron, and made
numerous bug fixes and improvements to the codebase.  This thing
is teetering on the verge of being useful.

Main chore ahead of me is rewriting the options parser to eliminate
all the funkiness and make it work as expected.  Not the only chore,
but the main one.