Start on build options for Zitron
The basic pattern here is going to carry us through.  Anything which can
be usefully configured at build time gets a `b.option` to do that, and
whatever the default value is gets overriden by command line flages.  I
need to completely rewrite the command line processor, or nearly so, but
while that's a bit of a slog it's really no big deal.