Parser for "rule name impls"
This is the trickier half of a commit to add a facility which allows for
code to be separated from the rule it's attached to.  This will combine
nicely with the preprocessor to enable debugging and reporting to be
available at one's fingertips.

The second half is adding an %impl directive which accepts a codeblock.
The grammar blocks code from being added to the rule if an impl
definition is present, and we'll allow multiples, so the fixup is a
simple matter of reporting orphan impls and welding the code on.