Clean exit
Zig install steps, quite reasonably, will not occur if a depended
process exits with an error.  Since inspecting the .out file is,
along with thinking Really Hard, the only resort in the event of
merge conflicts, this can be an issue during development.

As such, I've added a flag which exits zero under any circumstances.
Offer not valid for weird stuff like OOMs.